Load and format data object

Load and format data object

Usage

load_data(
  time,
  cluster_id,
  individual_id = NULL,
  treatment,
  outcome,
  time_type = "discrete",
  data
)

Arguments

time

A character string; the name of the numeric variable representing time. Time can be either discrete or continuous.

cluster_id

A character string; the name of the numeric variable identifying the cluster.

individual_id

A character string (optional); the name of the numeric variable identifying the individual.

treatment

A character string; the name of the binary variable indicating treatment. Values must be either integers (0/1) or Boolean (T/F).

outcome

Either a character string or a vector of two character strings; for a numeric or binary outcome, the single character string indicates the name of the numeric or binary outcome variable; for binomial outcome data, the vector of two character strings indicates the "# of successes" variable and the "# of trials" variable, respectively. Values in the outcome variable(s) must be either numeric or Boolean (T/F).

time_type

One of c("discrete", "continuous"); whether the model treats time as discrete or continuous.

data

A dataframe containing the stepped wedge trial data.

Value

An object of class sw_data

Examples

example_data <- load_data(time ="period", cluster_id = "cluster", individual_id = NULL,
treatment = "trt", outcome = "outcome_cont", data = sw_data_example)
#> Stepped wedge dataset loaded. Discrete time design with 15 clusters, 5 sequences, and 6 time points.
base::summary(example_data)
#>     outcome               time         cluster_id     treatment     
#>  Min.   :-3.440186   Min.   :1.000   4      : 153   Min.   :0.0000  
#>  1st Qu.:-0.714369   1st Qu.:2.000   13     : 146   1st Qu.:0.0000  
#>  Median : 0.028862   Median :3.000   6      : 139   Median :0.0000  
#>  Mean   : 0.008113   Mean   :3.437   7      : 138   Mean   :0.4863  
#>  3rd Qu.: 0.700841   3rd Qu.:5.000   11     : 137   3rd Qu.:1.0000  
#>  Max.   : 3.959173   Max.   :6.000   10     : 136   Max.   :1.0000  
#>                                      (Other):1162                   
#>   time_type         first_exposure exposure_time  
#>  Length:2011        Min.   :2.00   Min.   :0.000  
#>  Class :character   1st Qu.:3.00   1st Qu.:0.000  
#>  Mode  :character   Median :4.00   Median :0.000  
#>                     Mean   :3.99   Mean   :1.141  
#>                     3rd Qu.:5.00   3rd Qu.:2.000  
#>                     Max.   :6.00   Max.   :5.000  
#>